Search component - KakaoMap component 연결
